RavenDB-6783 Refactor Put method to a standalone class. Refactor ShouldRecreateAttachment method out of the Put method. Refactor identities related code to IdentitiesStorage. Move AddConflict and DeleteConflicts to ConflictsStorage. Remove duplicate GetChangeVectorEntriesFromTableValueReader method from ReplicationUtils.
