+++ test/FastTests/Server/Documents/DocumentsCrud.cs
@@ -556,8 +556,6 @@ namespace FastTests.Server.Documents
                 }
                 ctx.Transaction.Commit();
             }
-
-
         }
 
         public override void Dispose()
